refactor: add userpilot hook
